+++ R/mod_eda_select.R
@@ -7,19 +7,18 @@
 #' @noRd
 #'
 #' @importFrom shiny NS tagList
-mod_eda_freq_table_ui <- function(
-    id = NULL,
-    # --- Grouping
-    grouping_title = "Group data",
-    grouping_width = 12,
-    grouping_button_label = "Add group-by statement",
-    grouping_button_class = "btn-success",
-    grouping_button_style = "color: #fff;",
-    grouping_button_icon = icon('plus'),
-    grouping_button_width = 190,
-    # --- Freq table
-    freq_title = "Frequency table",
-    freq_width = 12
+mod_eda_select_ui <- function(
+    id = "eda_select",
+    select_title = "Select columns",
+    select_width = 12,
+    select_id = "select_ui",
+    select_button_label = "Add select statement",
+    select_button_class = "btn-success",
+    select_button_style = "color: #fff;",
+    select_button_icon = icon('plus'),
+    select_button_width = 170,
+    data_title = "Data table",
+    data_width = 12
 ) {
     ns <- NS(id)
 
@@ -29,34 +28,33 @@ mod_eda_freq_table_ui <- function(
         fluidRow(
             # column(
             shinydashboardPlus::box(
-                title = grouping_title,
-                width = grouping_width,
+                title = select_title,
+                width = select_width,
                 collapsible = TRUE,
                 actionButton(
-                    ns("add_group_by"),
-                    label = grouping_button_label,
-                    class = grouping_button_class,
-                    style = grouping_button_style,
-                    icon = grouping_button_icon,
-                    width = grouping_button_width
+                    ns("add_select"),
+                    label = select_button_label,
+                    class = select_button_class,
+                    style = select_button_style,
+                    icon = select_button_icon,
+                    width = select_button_width
                 ),
                 tags$br(),
                 tags$br(),
-                uiOutput(ns("grouping_ui"))
+                uiOutput(ns("select_ui"))
             )
         ),
         fluidRow(
             # column(
             shinydashboardPlus::box(
-                title = freq_title,
-                width = freq_width,
+                width = select_width,
+                title = select_title,
                 collapsible = TRUE,
-                DT::DTOutput(ns("grouping_tbl"))
+                DT::DTOutput(ns("select_tbl"))
             )
         ),
-        # sortable::sortable_js(ns("group_by_ui")),
         tags$script(src = "shimo.eda.js"),
-        tags$script(paste0("shimo_eda_mod_freq_table_js('", ns(''), "')"))
+        tags$script(paste0("shimo_eda_mod_select_js('", ns(''), "')"))
     )
 }
 
@@ -65,36 +63,36 @@ mod_eda_freq_table_ui <- function(
 #' @param id [[character]] Module ID
 #'
 #' @noRd
-mod_eda_freq_table_server <- function(
-    id = NULL,
-    r_data
+mod_eda_select_server <- function(
+    id = "eda_select",
+    r_data,
+    input_id_prefix = "select_input"
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
+        # browser()
 
-        # output$sortable_group_by_ui <- renderPrint({
-        #     input$group_by_ui
-        # })
-
-        # --- Create UI ---
-        input_ids <- get_input_ids(input_id_prefix = "grouping_input", sort = TRUE)
+        # --- Create select UI ----
+        input_ids <- get_input_ids(input_id_prefix = input_id_prefix, sort = TRUE)
         input_values <- get_input_values(input_ids = input_ids)
 
-        create_group_by_ui <- create_group_by_ui(
+        create_select_ui <- create_select_ui(
             r_data = r_data,
             input_ids = input_ids,
-            input_values = input_values
+            input_values = input_values,
+            input_id_prefix = input_id_prefix
         )
 
-        render_grouping_ui(
-            create_ui = create_group_by_ui
-        )
+        # output$select_ui <- renderUI({
+        #     create_select_ui()
+        # })
+        render_select_ui(id = NULL, create_select_ui = create_select_ui)
 
-        # --- Remove UI ---
-        remove_grouping_ui(id = NULL)
+        # --- Remove select UI ---
+        remove_select_ui(id = NULL)
 
         # --- Render data table ---
-        render_grouping_data_table(
+        render_select_data_table(
             id = NULL,
             r_data = r_data,
             input_ids = input_ids,
@@ -105,26 +103,27 @@ mod_eda_freq_table_server <- function(
 
 # Create inputs -----------------------------------------------------------
 
-create_group_by_ui <- function(
+create_select_ui <- function(
     id = NULL,
     r_data,
     input_ids,
     input_values,
-    input_id_prefix = "grouping_input"
+    input_id_prefix = "select_input",
+    input_id_button = "add_select"
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
+        input_id_pattern <- "^{input_id_prefix}" %>% stringr::str_glue()
 
-        eventReactive(input$add_group_by, {
+        eventReactive(input[[input_id_button]], {
             # browser()
-            cols <- r_data() %>% names()
-
             # --- Get input IDs and values ---
             input_ids <- input_ids()
             input_values <- input_values()
 
             # --- Handle existing select inputs ---
-            select_existing <- handle_existing_grouping_inputs(
+            cols <- r_data() %>% names()
+            select_existing <- handle_existing_select_inputs(
                 cols = cols,
                 input_values = input_values
             )
@@ -136,8 +135,10 @@ create_group_by_ui <- function(
             pos_latest <- length(select_existing) + 1
 
             input_id <- ns("{input_id_prefix}_{pos_latest}" %>% stringr::str_glue())
+            # print(input_id)
 
             select_new <- div(
+                id = input_id,
                 div(style = "display:inline-block; vertical-align:top; width:200px;",
                     selectInput(
                         inputId = input_id,
@@ -150,42 +151,13 @@ create_group_by_ui <- function(
                     actionButton(
                         inputId = ns("del_{input_id_prefix}_{pos_latest}" %>% stringr::str_glue()),
                         label = NULL,
-                        # icon = icon("times"),
                         icon = icon("trash-alt"),
                         width = 40,
                         class = "btn-danger delete_btn",
-                        # `data-toggle` = "tooltip",
-                        # `data-placement` = "top",
                         title = "Delete"
                     )
-                ),
-                id =input_id
+                )
             )
-
-            # id_del_button <- ns("del_{input_id_prefix}_{pos_latest}" %>% stringr::str_glue())
-            # --- Approach 1
-            # Via plain vanilla JS
-            # -> yes
-
-            # --- Approach 2
-            # shinyjs::onclick(id_del_button,
-            #     shinyjs::html(ns("time"), date()))
-            # shinyjs::onclick(id_del_button,
-            #     shinyjs::html(ns("ui_to_delete_id"), input_id))
-            # -> no
-
-            # --- Approach 3
-            # ns_prefix <- ns('')
-            # shinyjs::onclick(id_del_button,
-            #     shinyjs::runjs('Shiny.setInputValue(ns_prefix + "ui_to_delete_id", input_id, { priority: "event"});')
-            # )
-            # -> no
-
-            # --- Approach 4
-            # shinyjs::onclick(id_del_button,
-            #     updateTextInput(inputId = ns("ui_to_delete_id"), value = input_id)
-            # )
-            # -> yes
 
             tagList(
                 select_existing,
@@ -197,15 +169,16 @@ create_group_by_ui <- function(
 
 # Remove inputs -----------------------------------------------------------
 
-remove_grouping_ui <- function(
-    id = NULL
+remove_select_ui <- function(
+    id = NULL,
+    id_input = "ui_to_delete_id"
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
 
-        observeEvent(input$ui_to_delete_id, {
+        observeEvent(input[[id_input]], {
             # browser()
-            button_id <- input$ui_to_delete_id
+            button_id <- input[[id_input]]
 
             if (length(button_id) &&
                     button_id != ""
@@ -214,19 +187,11 @@ remove_grouping_ui <- function(
                     button_prefix = "del_"
                 )
 
-                # --- Approach 1: imperative removal
-                # if (select_id %in% names(ui_existing)) {
-                #     ui_existing[select_id] <- NULL
-                # }
-                # -> nope
-
-                # --- Approach 2: declarative removal
                 removeUI(
                     selector = "#{select_id}" %>% stringr::str_glue()
                 )
                 shinyjs::runjs('Shiny.onInputChange("{select_id}", null)' %>%
                         stringr::str_glue())
-                # -> yes
             }
         }, ignoreInit = TRUE)
     })
@@ -234,11 +199,11 @@ remove_grouping_ui <- function(
 
 # Handle existing inputs --------------------------------------------------
 
-handle_existing_grouping_inputs <- function(
+handle_existing_select_inputs <- function(
     id = NULL,
     cols,
     input_values = list(),
-    input_id_prefix = "grouping_input"
+    input_id_prefix = "select_input"
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
@@ -263,12 +228,9 @@ handle_existing_grouping_inputs <- function(
                             actionButton(
                                 inputId = ns("del_{input_id_prefix}_{.pos}" %>% stringr::str_glue()),
                                 label = NULL,
-                                # icon = icon("times"),
                                 icon = icon("trash-alt"),
                                 width = 40,
                                 class = "btn-danger delete_btn",
-                                # `data-toggle` = "tooltip",
-                                # `data-placement` = "top",
                                 title = "Delete"
                             )
                         )
@@ -283,41 +245,43 @@ handle_existing_grouping_inputs <- function(
 
 # Render UI ---------------------------------------------------------------
 
-render_grouping_ui <- function(
+render_select_ui <- function(
     id = NULL,
-    create_ui,
-    output_id = "grouping_ui"
+    create_select_ui,
+    output_id = "select_ui"
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
 
         output[[output_id]] <- renderUI({
-            create_ui()
+            create_select_ui()
         })
     })
 }
 
 # Render data table -------------------------------------------------------
 
-render_grouping_data_table <- function(
+render_select_data_table <- function(
     id = NULL,
     r_data,
     input_ids,
-    input_values,
-    output_id = "grouping_tbl"
+    input_values
 ) {
     shiny::moduleServer(id, function(input, output, session) {
         ns <- session$ns
 
-        output[[output_id]] <- DT::renderDT({
+        output$select_tbl <- DT::renderDT({
             data <- r_data()
 
-            input_ids <- input_ids()
-            if (length(input_ids)) {
+            group_by_ids <- input_ids()
+            if (length(group_by_ids)) {
                 cols <- input_values() %>% unname() %>% dplyr::syms()
                 if (length(cols)) {
+
+                    # data %>%
+                    #     wrang::wr_freq_table(!!!cols)
                     data %>%
-                        wrang::wr_freq_table(!!!cols)
+                        dplyr::select(!!!cols)
                 } else {
                     data
                 }