+++ dev/02_dev.R
@@ -18,8 +18,8 @@
 attachment::att_amend_desc()
 
 ## Add modules ----
-## Create a module infrastructure in R/
 golem::add_module(name = "eda_freq_table", with_test = TRUE)
+golem::add_module(name = "eda_select", with_test = TRUE)
 
 ## Add helper functions ----
 ## Creates fct_* and utils_*