+++ dev/app/server.R
@@ -1,10 +1,11 @@
 my_server <- function(input, output, session) {
     r_mtcars <- reactive({mtcars})
 
-    mod_eda_freq_table_server(
-        id = NULL,
-        r_data = r_mtcars
-    )
+    # --- Freq table ---
+    mod_eda_freq_table_server(id = "grouping", r_data = r_mtcars)
+
+    # --- Select ---
+    mod_eda_select_server(id = "test", r_data = r_mtcars)
 }
 
 my_server