+++ dev/app/ui.R
@@ -6,7 +6,8 @@ my_ui <- fluidPage(
         h1("Dev application", align = 'center'),
         windowTitle = "Dev application"
     ),
-    mod_eda_freq_table_ui(id = NULL)
+    mod_eda_freq_table_ui(id = "grouping"),
+    mod_eda_select_ui(id = "test")
 )
 
 my_ui