+++ NEWS.md
@@ -1,3 +1,12 @@
+# shimo.eda 0.0.0.9001 (2022-05-14)
+
+mod_eda_select
+
+- Added module `mod_eda_select.R`
+- Refactored module `mod_eda_freq_table.R`
+
+----------
+
 # shimo.eda 0.0.0.9000 (2022-05-14)
 
 Initial commit